requirements.txt
profanityfilter
torch
transformers==2.3.0
discord-py[voice]
google-cloud-texttospeech
azure-cognitiveservices-speech
